# docs/OPENAPI.yaml
openapi: 3.1.0
info:
  title: FiPro_1 API
  version: 0.1.0
  description: Minimal API contract for MVP.
servers:
  - url: http://localhost:8000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: INVALID_ARGUMENT }
        message: { type: string }
        retryable: { type: boolean, default: false }
    GenerateReportRequest:
      type: object
      required: [ticker, asof, strategy_version_id, tier]
      properties:
        ticker: { type: string, example: 600519.SH }
        asof: { type: string, format: date-time }
        strategy_version_id: { type: string, example: stg_v1 }
        tier:
          type: string
          enum: [TIER0, TIER1, TIER2]
    GenerateReportResponse:
      type: object
      required: [report_id, final_report]
      properties:
        report_id: { type: string, example: rpt_20260210_600519 }
        final_report: { $ref: '#/components/schemas/Report' }
    Report:
      type: object
      required:
        - schema_version
        - report_id
        - generated_at
        - ticker
        - asof
        - strategy_version_id
        - tier
        - decision
        - price_bands
        - key_drivers_to_watch
        - thesis
        - risk_flags
        - invalidations
        - evidence_refs
        - data_quality
        - provenance
        - memory_update
      properties:
        schema_version: { type: string, example: "0.1" }
        report_id: { type: string }
        generated_at: { type: string, format: date-time }
        ticker: { type: string }
        market: { type: string, enum: [CN_A, US, HK, CRYPTO, OTHER] }
        asof: { type: string, format: date-time }
        strategy_version_id: { type: string }
        tier: { type: string, enum: [TIER0, TIER1, TIER2] }
        decision: { type: object }
        price_bands:
          type: array
          items: { type: object }
        key_drivers_to_watch:
          type: array
          items: { type: object }
        thesis: { type: object }
        risk_flags:
          type: array
          items: { type: object }
        invalidations:
          type: array
          items: { type: object }
        evidence_refs:
          type: array
          items:
            type: object
            required: [evidence_id, type, title, source, captured_at]
            properties:
              evidence_id: { type: string }
              type: { type: string, enum: [SNAPSHOT_FIELD, NEWS_DOC, FILINGS, MACRO_SERIES, GRAPH_QUERY, MANUAL_NOTE] }
              title: { type: string }
              source: { type: string }
              captured_at: { type: string, format: date-time }
              uri: { type: string }
              snippet: { type: string }
              checksum: { type: string }
        data_quality: { type: object }
        provenance: { type: object }
        memory_update: { type: object }
    WatchlistItem:
      type: object
      required: [ticker, tier]
      properties:
        ticker: { type: string }
        tier: { type: string, enum: [TIER0, TIER1, TIER2] }
    ReportFeedbackRequest:
      type: object
      required: [feedback_label]
      properties:
        feedback_label: { type: string, enum: [USEFUL, USELESS, FALSE_POSITIVE] }
        comment: { type: string, default: "" }
    ReportFeedbackItem:
      type: object
      required: [feedback_id, report_id, feedback_label, created_at]
      properties:
        feedback_id: { type: string }
        report_id: { type: string }
        feedback_label: { type: string, enum: [USEFUL, USELESS, FALSE_POSITIVE] }
        comment: { type: string }
        created_at: { type: string, format: date-time }
    ReportFeedbackListResponse:
      type: object
      required: [report_id, items]
      properties:
        report_id: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/ReportFeedbackItem' }
paths:
  /health:
    get:
      summary: Service health check
      responses:
        '200': { description: OK }
  /version:
    get:
      summary: Service version
      responses:
        '200': { description: OK }
  /strategies:
    post:
      summary: Create strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201': { description: Created }
        '400':
          description: Bad Request
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /strategies/{id}/versions:
    post:
      summary: Create strategy version
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '201': { description: Created }
  /strategies/{id}/versions/{version}:
    get:
      summary: Get strategy version detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: version
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /tickers/{ticker}/snapshot:
    get:
      summary: Get market snapshot
      parameters:
        - in: path
          name: ticker
          required: true
          schema: { type: string }
        - in: query
          name: asof
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: strategy_version
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /reports/generate:
    post:
      summary: Trigger report workflow
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerateReportRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateReportResponse' }
  /reports/{report_id}:
    get:
      summary: Get report detail
      parameters:
        - in: path
          name: report_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerateReportResponse' }
        '404':
          description: Not Found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /reports/{report_id}/feedback:
    post:
      summary: Submit report feedback label
      parameters:
        - in: path
          name: report_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportFeedbackRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportFeedbackItem' }
        '404':
          description: Not Found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      summary: List feedback entries for a report
      parameters:
        - in: path
          name: report_id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportFeedbackListResponse' }
  /watchlist:
    get:
      summary: Get watchlist
      responses:
        '200': { description: OK }
    post:
      summary: Add/update watchlist tier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WatchlistItem' }
      responses:
        '200': { description: OK }
  /graph/subtree:
    get:
      summary: Query supply chain subtree
      parameters:
        - in: query
          name: ticker
          required: true
          schema: { type: string }
        - in: query
          name: depth
          required: true
          schema: { type: integer, minimum: 1, maximum: 6 }
      responses:
        '200': { description: OK }
  /graph/paths:
    get:
      summary: Query impact paths
      parameters:
        - in: query
          name: entity
          required: true
          schema: { type: string }
        - in: query
          name: ticker
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /memory/search:
    get:
      summary: Search memory
      parameters:
        - in: query
          name: ticker
          required: true
          schema: { type: string }
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /memory/write:
    post:
      summary: Write memory note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ticker, content]
              properties:
                ticker: { type: string }
                content: { type: string }
                tags:
                  type: array
                  items: { type: string }
      responses:
        '201': { description: Created }